+++ DuckEngine/DuckEngine/include/DuckEditor/Layers/Layer.h
@@ -23,7 +23,7 @@ public:
 	Layer() = default;
 	virtual ~Layer() = default;
 
-	virtual void Update([[maybe_unused]] f64 dt) {}
+	virtual void Update([[maybe_unused]] f32 dt) {}
 
 	//Overwrite this!
 	static LayerType GetType() { return LayerType::Invalid; }
